+++ bostonBusMap/src/boston/Bus/Map/data/LocationComparator.java
@@ -5,14 +5,14 @@ import java.util.Comparator;
 import boston.Bus.Map.math.Geometry;
 
 public class LocationComparator implements Comparator<boston.Bus.Map.data.Location> {
-	private final float centerLatitudeAsRadians; 
-	private final float centerLongitudeAsRadians;
+	private final double centerLatitudeAsRadians; 
+	private final double centerLongitudeAsRadians;
 
 
-	public LocationComparator(float centerLatitude, float centerLongitude)
+	public LocationComparator(double centerLatitude, double centerLongitude)
 	{
-		centerLatitudeAsRadians = (float) (centerLatitude * Geometry.degreesToRadians);
-		centerLongitudeAsRadians = (float) (centerLongitude * Geometry.degreesToRadians);
+		centerLatitudeAsRadians = centerLatitude * Geometry.degreesToRadians;
+		centerLongitudeAsRadians = centerLongitude * Geometry.degreesToRadians;
 	}
 
 	public int compare(Location a, Location b)
