+++ bostonBusMap/src/boston/Bus/Map/transit/TransitSystem.java
@@ -157,8 +157,8 @@ public class TransitSystem {
 	}
 
 	public void refreshData(RouteConfig routeConfig,
-			int selectedBusPredictions, int maxStops, float centerLatitude,
-			float centerLongitude, ConcurrentHashMap<Integer, BusLocation> busMapping,
+			int selectedBusPredictions, int maxStops, double centerLatitude,
+			double centerLongitude, ConcurrentHashMap<Integer, BusLocation> busMapping,
 			String selectedRoute, RoutePool routePool,
 			Directions directions, Locations locations) throws IOException, ParserConfigurationException, SAXException {
 		for (TransitSource source : transitSources)
